Move news gallery view in shared
