+++ src/app/main/component/eco-news/components/index.ts
@@ -3,7 +3,7 @@ export { EcoNewsDetailComponent } from './eco-news-detail/eco-news-detail.compon
 export { EcoNewsWidgetComponent } from './eco-news-detail/eco-news-widget/eco-news-widget.component';
 export { NewsListComponent } from './news-list/news-list.component';
 export { ChangeViewButtonComponent } from './news-list/change-view-button/change-view-button.component';
-export { NewsListGalleryViewComponent } from './news-list/news-list-gallery-view/news-list-gallery-view.component';
+export { NewsListGalleryViewComponent } from 'src/app/shared/news-list-gallery-view/news-list-gallery-view.component';
 export { NewsListListViewComponent } from './news-list/news-list-list-view/news-list-list-view.component';
 export { NewsPreviewPageComponent } from './news-preview-page/news-preview-page.component';
 export { PostNewsLoaderComponent } from './post-news-loader/post-news-loader.component';
