+++ src/app/main/component/user/components/habit/add-new-habit/add-new-habit.component.html
@@ -72,5 +72,6 @@
       </button>
     </div>
   </div>
-  <app-habits-widget [tag]="tags[0]"></app-habits-widget>
+  <app-habits-widget [tag]="tags[0]" [isHabit]="true"></app-habits-widget>
+  <app-habits-widget [isHabit]="false"></app-habits-widget>
 </div>
