+++ src/app/main/component/user/components/habit/add-new-habit/habits-widget/habits-widget.component.html
@@ -1,13 +1,26 @@
 <div class="wrapper">
   <p>
-    {{ 'user.habit.habit-widget.title' | translate }}
+    {{ (isHabit ? 'user.habit.habit-widget.title' : 'user.habit.habit-widget.title-news') | translate }}
     <span>{{ tag }}</span>
   </p>
   <hr />
-  <div class="habits-recomended">
-    <div *ngIf="recommendedHabits.length" class="cards-container">
+  <div *ngIf="isHabit" class="habits-recomended">
+    <div class="cards-container">
       <div class="habit-card" *ngFor="let habit of recommendedHabits; index as i">
-        <app-habits-gallery-view [habit]="habit"></app-habits-gallery-view>
+        <app-habits-gallery-view [habit]="habit" [routerLink]="['/news', habit.id]"> </app-habits-gallery-view>
+      </div>
+    </div>
+  </div>
+  <div *ngIf="!isHabit" class="habits-recomended">
+    <div class="cards-container">
+      <div class="habit-card" *ngFor="let news of recommendedNews; index as i">
+        <app-news-list-gallery-view
+          [ecoNewsModel]="news"
+          [routerLink]="['/news', news.id]"
+          [class.last-item]="i === 2"
+          [class.recommended-item]="true"
+        >
+        </app-news-list-gallery-view>
       </div>
     </div>
   </div>
