+++ src/app/shared/news-list-gallery-view/news-list-gallery-view.component.html
