+++ src/assets/i18n/ua.json
@@ -745,7 +745,8 @@
         "great": "Чудово! Ви набули звичку!"
       },
       "habit-widget": {
-        "title": "Більше звичок пов'язаних з "
+        "title": "Більше звичок пов'язаних з ",
+        "title-news": "Перегляньте наші корисні посилання"
       }
     }
   },
